mongodb-gcp/iotcore_to_mongodb/requirements.txt
pymongo
dnspython
requests
flask
pyyaml